Updated openstack/openstack Project: openstack/horizon  4b300a46e58edae3d600f627d7d25e30498307b5
Change unit string translation in volume limit templates

This fixes an issue where the lazy translated "GB" unit string is
not properly resolved in the volume limit templates. This error
does not happen in a devstack environment but has been found to
happen in chef and puppet deployed OpenStack production
environments.

Closes-Bug: 1332238
Change-Id: Icc056ca508f430e981049ca00985d329c92a6ab4