Updated openstack/openstack Project: openstack/nova  aecaf3a14aee39c374f9aaf5fd30b61d8bb7388e
Require tests for DB migrations

This patch makes it mandatory to have at least a check routine for each migration.
Some migrations are special (like the compressed one at the beginning, and the
placeholder migrations), and can be skipped. Also, all baremetal migrations are
excepted from this policy since that code is deprecated and needs to go away soon.

Change-Id: Id8c10d5ea3d8a8864a32ea4b0f0b94657e1e9bf7
Closes-Bug: #1370536