Updated openstack/openstack Project: openstack/horizon  b25fb96796d488aefa5e08b190e91c1cc54297a6
Style network address for subnet as required field

It does this by adding the required class to the network address field when the
checkbox is enabled.

I have also made the checkbox behave like a switchable select so when the create
subnet checkbox is checked all fields are visible. There are hidden when create
subnet is unselected.

Change-Id: Id45d5ce2a1bb55e4bcaa33df065ca2b1ccee3c7e
Closes-bug: 1367858