Updated openstack/openstack Project: openstack/django_openstack_auth  455aaeab8d16fbecebe9a088e10e0e532a3116cb
Fix Django 1.7 compat

In Debian, we are going to release Jessie with Django 1.7,
which has just been released. Therefore, compatibility is
important for us.

This patch, which the Debian package already carries, fixes
2 small compatibility issues with Django 1.7. It is safe to
use openstack_auth with these patches and a lower version of
Django, which makes it safe to apply this patch.

Change-Id: I564cde889c56e12fc0fc1347537e66b576605a0d