Updated openstack/openstack Project: openstack/keystone  b2c50822a6df904d8f8dffc1948bd51841bfab75
Set the default number of workers when running under eventlet

Keystone now defaults to n-cpu as the number of workers under eventlet with
a minimum number of workers when auto-calculating the number of 2.

This behavior can be overridded by changing the `admin_workers` and
`public_workers` configuration values in the Keystone config file.

Change-Id: I482cfef7ee14e995c6fb206bcef412957c54f491
Closes-Bug: #1371154