Updated openstack/openstack Project: openstack/nova  d95f767c670e021c5ebb7cc466f55e2ec665f902
Don't refresh network cache for instances building or deleting

Refreshing the network cache for an instance which is still building
is both unnecessary (the build itself will update the cache when it
completes) and can lead to a race condition if the periodic task gets
incomplete network information, then gets pre-empted by the build
thread, and then updates the cache after the build has finished.

Likewise there is no value in updating the network cache for an
instance which is in the process of being deleted.

Closes-Bug: 1254320
(cherry picked from commit 713e538237c7fea0b93ade343b1d9368bdbf2698)

Conflicts:
	nova/compute/manager.py

Change-Id: I7dd1d0a005662b979eaea3476af2506cbc51f17a