Updated openstack/openstack Project: openstack/requirements  43fa3d0221d408580cdd03c2dddbdba11cecd7f6
Upgrade PyScss to version 1.2.1 and pin it <1.3 for Horizon

PyScss fixed an important bug in parsing Bootstrap 3.2.0
(https://github.com/Kronuz/pyScss/issues/287). While investigating
this book with the upstream developers, we also became aware that
version 1.3 will break backwards compatibility with Django-PyScss
that we are using, so I'm pinning the version.

Change-Id: Ic4c23dcbf35dd55916401bba2b1bc3bf4baa4522
Partial-bug: #1370137