Updated openstack/openstack Project: openstack/nova  fdcf6b8c79e1fecb998ff40e8330b2ad50fb3b7d
XenAPI improve post snapshot coalesce detection

The coalesce detection is not working well after this change:
ae2a27ce19f3e24d4a8c713a73e617f4cd71d4b4

The snapshot operation will introduce a new VHD file into the VDI chain,
and in many cases, that would be coalesced during the next SR scan. So
we need to walk the chain after the snapshot has been taken, not before.

This fixes the cause of the mentioned bug, but it doesn't help with
launching the partially corrupted snapshots created because of this bug,
so this only partially fixes the bug.

Change-Id: I0bf7535ae3f7d9e6820f8dc07075892953d80a78
Partial-Bug: #1362595