Updated openstack/openstack Project: openstack/neutron  52c967aeebad4ab53d9d8efb2f580b3dae117e51
Indicate the begin and end of the sync process to EOS

Send a trigger to EOS when a sync operation is initiated, and,
another trigger when the sync operation is complete.
Additionally, sync_interval value (from ml2_conf_arista.ini)
is passed down to EOS. This is used by EOS to timeout the
transaction.

Change-Id: I5110b24380ad848d03719e94283d69a7bc0ce972
Closes-Bug: 1371324