Updated openstack/openstack Project: openstack/python-ceilometerclient  de7affee29f0412e68f3642c7ed3e86f75a7f993
Reduce redundant parameter of some commands in CLI

When deleting an alarm, we use "ceilometer alarm-delete -a <ALARM_ID>",
unlike other deleting commands of openstack, the parameter-a/--alarm_id
is redundant. The similar situations exist in showing alarm, geting alarm
state, showing resource and so on.
It is more easy to use for reducing these parameters.

New behaviour:

$ ceilometer help alarm-show
usage: ceilometer alarm-show [<ALARM_ID>]

Show an alarm.

Positional arguments:
  <ALARM_ID>  ID of the alarm to show.

$ ceilometer alarm-show
alarm_id should not be empty

$ ceilometer alarm-show abcde
Not Found (HTTP 404)

$ ceilometer alarm-show -a abcde
-a is obsolete! See help for more details.
Not Found (HTTP 404)

$ ceilometer alarm-show --alarm_id abcde
--alarm_id is obsolete! See help for more details.
Not Found (HTTP 404)

Co-Authored-By: Nejc Saje <nsaje@redhat.com>
Change-Id: I1fbc85aa253929bfbb5e73ed834a725b9cf828b4
Closes-bug: #1268557