Updated openstack/openstack Project: openstack/tempest  26fa3d72d47edfd6efda0a1f2b253d62342b4b7c
Fix unstable assertion in test_cinder_endpoints

In simple_read_only.test_cinder.test_cinder_endpoints, the test case
assumes that the first table of 'cinder endpoints' always ['nova',
'Value']. However, it is sometimes wrong.
So in this commit, the test verifies that they have 2 or more headers
and has a 'Value' header.

Change-Id: Idb3f2ebc6b7e05178621ac3701b7c3d02e1d7591
Closes-Bug: #1363018
Closes-Bug: #1362703