Updated openstack/openstack Project: openstack/horizon  25855683c49735060c246dc3114dc60ab4794c62
Ability to add form confirm/cancel button labels

We repeat a lot of templating code in our forms. It makes sense to abstract
some of these repetitive logic and reduce the clutter. This is an unintrusive
patch aim to give users ability to specify confirm and cancel labels in the
form class.

Change-Id: I6eba93c61a3a6a94ac7b42f199bfb95df86e96b6
Partially-Implements: blueprint form-template-to-view