Updated openstack/openstack Project: openstack/nova  19587df028ce8fb27f30a845c880505398f75b9a
Fix 'force' parameter for quota-update

Commit 77b4012a02c6b4827a7a6b112f35deb9ce94c395 added further
validation on the quota update controller that isn't aware of the
'force' parameter. This validation also duplicated the existing
validation that *did* honour 'force'.

This change merges the two validation methods and ensures that the
validation is skipped when the 'force' parameter is passed in the
request.

Change-Id: I25e3ce857a83cb50789a2a61bb01c796128ff4f8
Closes-bug: #1362405