Updated openstack/openstack Project: openstack/keystonemiddleware  1f8b4fe4432c7725ed3ac630ab07990ff9a48d17
Fix test failure after discovery hack

With the discovery hack, the test_auth_plugin is now trying to
query the unversioned endpoint from the catalog to discover what
versions it supports. Since that URL wasn't stubbed out with
httpretty it was causing a failure and then a different URL than
expected was returned by the plugin.

This change registers the discovery URL that's in the test token
so that fetching it doesn't fail but returns a normal version
response.

Related-Bug: #1335726
Closes-Bug: #1372190
Change-Id: I1ae41801ed44ec6a37d972f62fc853dfd609128f