Updated openstack/openstack Project: openstack/swift  607e2ed2e30ca02d114d04484763d92afb9158ab
Add "--no-overlap" option to swift-dispersion populate

This change allows the user to use a "--no-overlap" parameter when
running the tool multiple times. It will increase the coverage by
whatever is specified in the dispersion_coverage field of the conf
file in a manner where existing container/objects are left in place
and no partition is populated more than once.

Related-Bug: #1233045

Change-Id: I139fed2f4c967ba18d073b7ecd1e946ed4da1271