Updated openstack/openstack Project: openstack/taskflow  5932e2b865f6ea719e63c569bdf16bb1fc9d84e2
LOG which requeue filter callback failed

When a requeue filter fails being called instead of just
logging the exception and traceback also log the callback
index which failed and the string representation of the
callback that failed so that when/if it is later examined to
determine the root cause these pieces of output can be used
to isolate the failing callback.

Also includes a drive-by commit to update the add requeue
filter callback documentation to explain what the callback
will be provided and expected to return.

Change-Id: Idf1ae63cfa706551ac0361e802f3b082783e0b9d