Updated openstack/openstack Project: openstack-infra/devstack-gate  5eeb0bcabdf6807682df0264c8d842358ae68efd
Disable cinder with ironic

Cinder doesn't currently work with ironic as volumes cannot yet
be attached to baremetal nodes. This is something that may be
available in the K series. For now, though, so that we can work
on enabling the tempest API and scenario tests with ironic, we
need to disable cinder.

Change-Id: I783d3c8539acc2e74922a8aad98f0c7928b3815e