Updated openstack/openstack Project: openstack/nova  51de439a4d1fe5e17d59d3aac3fd2c49556e641b
make flavors use common limit and marker

Flavors REST calls were not using the common limit and 
marker function, which means that the osapi_max_limit configuration
that was set by the operator would not be respected. 

Convert flavors to use the common infrastructure thereby providing
a more consistent experience for users and operators.

Change-Id: Iabbf74cf14ddab9a67092ba966270a658ce6be3f
Closes-Bug: #1304790