Updated openstack/openstack Project: openstack-infra/devstack-gate  068c3ed3c06f519b57ec53d9d8d5526abaa9e0ea
Switch to static content for glance upload test

glance is able to fetch an image from a given url.
Previously the dashboard main page was used as image,
this change sets it a static image, in order
the have less gate issue.

Any url could be used which does not requires authentication.

Change-Id: I66760790aa7d2d0d7b8c53d62baf59134b9b44d6