Updated openstack/openstack Project: openstack/tempest  64fc93510ec3208f2ded961dd9c50d8c648f6ced
Fix scenarios not passing down specific network

Nova v2 api expects nics in specific networks to be passed
down in an array of "networks". nics is used in novaclient
but converted to networks in
python-novaclient/tree/novaclient/v1_1/servers.py#n520

Apply the conversion to the parameters to make the check
pass in case where there would be multiple networks
matching the requirements. Also remove a case where scenario
manager forced in the wrong network into the tests.

Closes-Bug: #1368834
Change-Id: Ic0b87377c5878e240a0fbcc6549023c828dd97ab