Updated openstack/openstack Project: openstack/requirements  1508a274ed2bee9e3efb8d84c3a93dfbdf1c5c4f
Add a tox env to check requirements overlap

There was no easy way to check that the requirements in the
current commit overlap with requirements in the ancestor commit.

This change adds a tox env that when run compares the
global-requirements.txt with the parent commit's
global-requirements.txt. If the requirements for the packages don't
overlap it fails.

Run it with

 tox -e versions-overlap-parent

Change-Id: I03bdead39760e9c13d6de4a33f8e5e788052323d
Closes-Bug: #1326504