Updated openstack/openstack Project: openstack/horizon  4c3d12dbdb6848173bd0447a28b6122ff0b7be70
Fixes double evaluation of AngularJS in modal box

AngularJS syntax is evaluated twice in modal box after submitting form
which succeed and no redirection is going to be made. This commit fixes
this bevahiour by executing AngularJS code just after modal is created
and not every time it is going to be shown.

Change-Id: I204fce23ce52f74c38b7dc40ef03c6d5710ad254
Closes-Bug: #1371116