Updated openstack/openstack Project: openstack/horizon  9f426ec6731874627cf8615d9f8ef23125bf77cb
Add mixin to replace replicated get_policy_target

Fix adds a mixin to replace the replicated get_policy_target
function. Duplicated code is removed from many classes and mixin is
inherited.

project_id, user_id and domain_id are now default in the generated
policy_target dict but are set to None if the data doesn't support it,
this is to provide the most flexibility for operators writing policy.json
as discussed in the bug report.

policy_target_attrs attribute added by the mixin is overwrittable by
sub-classes of the mixin to override the defaults for custom policy_target
information.

Change-Id: I26759f145b8756bd1eef585c8107160277061523
Closes-Bug: 1317238