Updated openstack/openstack Project: openstack/keystonemiddleware  f56e5fef3a6121079f745188dc4f64f9b707477b
Fix auth_token for old oslo.config

When running with a havana-level of oslo.config (<1.3.0),
applications with any config options in their api-paste.ini will
fail to start with an error like

 'StrOpt' object has no attribute 'type'

This is because the config options didn't have a type attribute
until 1.3.0.

During the grenade test, the havana level of oslo.config is used.

Change-Id: I745c3e04f18941a2d41e191d43f61b926522bb9d
Closes-Bug: #1372422