Updated openstack/openstack Project: openstack/oslo-incubator  5d40e1431ea0f42d088e44d776264b141ac0947b
Remove code that moved to oslo.i18n

Remove the modules that went to oslo.i18n and their tests, then update
the other modules and tests to use oslo.i18n.

There are a few cases where other modules that will be deleted from the
incubator for juno had references to gettextutils, especially for their
tests. These are commented out as an expedient way to make the tests
run, in anticipation of the files being removed separately.

Change-Id: Ib80a3f0b0f6e96065e86a9b477aebd8bce1201ba
Closes-Bug: #1371702