Updated openstack/openstack Project: openstack/keystone  d8d1477d83f294edb758d22e08aee919f4cf6d75
Prevent infinite recursion on persistence core on init

Raise an AttributeError for lookup on 'token_provider_api' within the
__getattr__ on the keystone.token.persistence.core.Manager class. The
__getattr__ will be called on dependency injection process dependencies
and attempt to do a lookup on .token_provider_api, which nets an
infinite recursion. If an AttributeError is raised the dependency
processor will set the value correctly. Once the 'token_provider_api'
attribute is set, __getattr__ is no longer called as the value is in
the instantiated object's __dict__.

Change-Id: I4af79f4568429dea27c1e11e5a77cefa5afc792c
Closes-Bug: #1373167