Updated openstack/openstack Project: openstack/keystone  15780e895fe532fb8e1120db3160f21e8d154409
Adding an index on token.user_id and token.trust_id

This is a performance update to ensure that we are scanning the fewest
number of rows on a user delete (causing token revocations). Without
these indexes it is possible to scan all valid tokens, causing
significant overhead, to find the user or trust matching tokens.

Due to selecting the extra column (needed for other matches in
some cases) this can also cause issues with buffer pool sizes.

Change-Id: I202b5c87a221d8dba99d16b0a1baa7546fef093b
Closes-Bug: 1332666