Updated openstack/openstack Project: openstack/keystone  4b917f9c6fb9f110bb1f70fa4785fde5999a6dac
Fix failure of delete domain group grant when identity is LDAP.

When deleting a domain group assignment while using a non
domain-aware backend, such as LDAP, an AttributeError was being
raised when trying to find all the relevent tokens. This was due
to a hang over from when you had to pass domain scope to
list_user_in_group(). This only affected domain group grants,
by luck we got away with it for group project grants.

Change-Id: I47b61886698232a7d3dfb4b502d61723cb0eb786
Closes-Bug: 1373113