Updated openstack/openstack Project: openstack/heat  9fdbf468248e0755ac6414a68a47e3a79bdf5cc2
Resolve AWS::EC2::Instance AZ output to a value if not specified

The availability zone output attribute of the instance resource is
empty if no input is provided by the user for that property.
Instead of relying on solely user input to determine the value for this,
this attribute must be resolved to an appropriate value.

The VolumeTest.test_volume_default_az test case had an unmocked function
which is now mocked here to successfuly test this feature.

Change-Id: Iab2cf1f3da0261c30b755d0b87ebb693644a4a29
Closes-Bug: #1334519