Updated openstack/openstack Project: openstack/tempest  736c9c554b8c9773f4def7350a3feb621e4170fb
Properly preserve trace on error during wait

This commit starts using save_and_rereraise_exception from oslo
excutils when handling an server ERROR state during a wait loop.
Previously, if the wait loop in the common api create_server method
caught a BuildErrorException it attempts to do some final processing
and cleanup around that server. However, doing this interferes with
the trace output from the failure and only would show the subsequent
re-raise as the stack trace. This commit avoids that but using the
oslo method to get around this.

Change-Id: I62193af963020311d5ed90a847c9764faff3e4c2